+++ lib/src/parent_isolate.dart
@@ -6,9 +6,7 @@ import "package:typed_isolate/typed_isolate.dart";
 import "package:burt_network/burt_network.dart";
 import "package:video/src/collection.dart";
 import "package:video/src/frame.dart";
-import "package:video/src/functions.dart";
 
-import "server.dart";
 
 class VideoController extends IsolateParent<VideoCommand, FrameData>{
 
@@ -22,6 +20,7 @@ class VideoController extends IsolateParent<VideoCommand, FrameData>{
       collection.videoServer.sendMessage(VideoData(frame: frame.data, details: data.details));
       frame.dispose();
     } else {
+      print(data.details);
       collection.videoServer.sendMessage(VideoData(details: data.details));
     }
   }
