+++ lib/video.dart
@@ -1,4 +1,4 @@
-export "src/camera.dart";
+export "src/camera_isolate.dart";
 export "src/collection.dart";
 export "src/constants.dart";
 export "src/server.dart";
